src/main/java/Figure/Point.java
package Figure;

public class Point {
    private int x, y;

    Point(int x, int y) {
        this.x = x;
        this.y = y;
    }

    private int getX() {
        return x;
    }

    private int getY() {
        return y;
    }

    public void setX(int x) {
        this.x = x;
    }

    public void setY(int y) {
        this.y = y;
    }

    public static double getDistance(Point p1, Point p2) {
        return Math.sqrt(Math.pow(p2.getX() - p1.getX(), 2) + Math.pow(p2.getY() - p1.getY(), 2));
    }

    public static Point getThirdPoint(Point p1, Point p2) {
        return new Point(p1.getX(), p2.getY());
    }

    public String toString() {
        return getX() + " " + getY();
    }
}
